backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
